+++ README.dev.md
@@ -40,6 +40,7 @@ Geospatial
 ```
 uv run dart-pipeline get meteorological/aphrodite-daily-mean-temp l=INFO unpack
 uv run dart-pipeline get geospatial/gadm 3=VNM
+uv run dart-pipeline process geospatial/aphrodite-daily-mean-temp 3=VNM a=0 d=2015-05-11 l=INFO
 uv run dart-pipeline process geospatial/aphrodite-daily-mean-temp 3=VNM a=0 d=2015-05-11 l=INFO plots
 uv run dart-pipeline process geospatial/aphrodite-daily-mean-temp 3=VNM a=1 d=2015-05-11 l=INFO plots
 uv run dart-pipeline process geospatial/aphrodite-daily-mean-temp 3=VNM a=3 d=2015-05-11 l=INFO plots
@@ -50,6 +51,7 @@ uv run dart-pipeline process geospatial/aphrodite-daily-mean-temp 3=VNM a=3 d=20
 ```
 uv run dart-pipeline get meteorological/aphrodite-daily-precip -u
 uv run dart-pipeline get geospatial/gadm 3=VNM
+uv run dart-pipeline process geospatial/aphrodite-daily-precip 3=VNM a=0 d=2015-05-11 l=INFO
 uv run dart-pipeline process geospatial/aphrodite-daily-precip 3=VNM a=0 d=2015-05-11 l=INFO plots
 uv run dart-pipeline process geospatial/aphrodite-daily-precip 3=VNM a=1 d=2015-05-11 l=INFO plots
 uv run dart-pipeline process geospatial/aphrodite-daily-precip 3=VNM a=3 d=2015-05-11 l=INFO plots
